UPD u8 result

+++ src/dice_roller/roll_result.rs
@@ -1,13 +1,13 @@
 /// A struct holding dice results
 #[derive(Debug)]
 pub struct RollResult {
-    pub dice: Vec<i32>,
+    pub dice: Vec<u8>,
     pub successes: Vec<i32>,
     pub outcome: i32,
 }
 
 impl RollResult {
-    pub fn new(dice: Vec<i32>, outcome: i32) -> Self {
+    pub fn new(dice: Vec<u8>, outcome: i32) -> Self {
         Self {
             dice,
             outcome,
